requirements.txt
python-telegram-bot==21.3
openai==1.40.0
apscheduler==3.10.4
python-dotenv==1.0.1
